feat: gameplay instructions
-Add gameplay instructions for classic mode & display it underneath the virtual keyboard.

-Refactor main gameplay to a new component for reusability.

-Add logic to redirect the info icon link in the top menu to navigate to the gameplay instructions section of the page when clicked.
